requirements.txt
python-telegram-bot==20.0
gspread==5.12.0
python-dotenv==1.0.1
oauth2client==4.1.3
matplotlib